# v1EUOpenAPI.json
{
    "openapi": "3.0.3",
    "info": {
        "version": "v1",
        "title": "European Parliament Open Data API",
        "description": "API giving access to data published on the Open Data Portal of the European Parliament (EP).",
        "license": {
            "name": "European Parliament Legal Notice",
            "url": "https://www.europarl.europa.eu/legal-notice"
        }
    },
    "servers": [
        {
            "url": "https://data.europarl.europa.eu/api/v1",
            "description": "EP Open Data API"
        }
    ],
...
    "components": {
        "schemas": {
            "MEPS": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "required": [
                                "id",
                                "type"
                            ],
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "sortLabel": {
                                    "type": "string"
                                },
                                "identifier": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "givenName": {
                                    "type": "string"
                                },
                                "familyName": {
                                    "type": "string"
                                },
                                "officialGivenName": {
                                    "type": "string"
                                },
                                "officialFamilyName": {
                                    "type": "string"
                                },
                                "api:country-of-representation": {
                                    "type": "string"
                                },
                                "api:political-group": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "parameters": {
...
            "format-all": {
                "name": "format",
                "in": "query",
                "description": "Response format",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "application/ld+json",
                    "enum": [
                        "application/ld+json",
                        "application/rdf+xml",
                        "text/turtle"
                    ]
                }
            },
...
            "gender": {
                "name": "gender",
                "in": "query",
                "description": "Gender of the person. The non-exhaustive list of values includes concepts of the Human sex authority table maintained by the Publications Office of the European Union: https://op.europa.eu/en/web/eu-vocabularies/concept-scheme/-/resource?uri=http://publications.europa.eu/resource/authority/human-sex .",
                "required": false,
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "FEMALE",
                            "MALE",
                            "NAP",
                            "NKN",
                            "NST"
                        ]
                    }
                }
            },
            "country-of-representation": {
                "name": "country-of-representation",
                "in": "query",
                "description": "27 EU Member States and UK. Values are ISO 3166-1 alpha-2 country codes.",
                "required": false,
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "BE",
                            "BG",
                            "CZ",
                            "DK",
                            "DE",
                            "EE",
                            "IE",
                            "EL",
                            "ES",
                            "FR",
                            "HR",
                            "IT",
                            "CY",
                            "LV",
                            "LT",
                            "LU",
                            "HU",
                            "MT",
                            "NL",
                            "AT",
                            "PL",
                            "PT",
                            "RO",
                            "SI",
                            "SK",
                            "FI",
                            "SE",
                            "UK"
                        ]
                    }
                }
            },
            "political-group": {
                "name": "political-group",
                "in": "query",
                "description": "EP political groups. The non-exhaustive list of values includes  concepts of the Corporate body authority table maintained by the Publications Office of the European Union: https://op.europa.eu/en/web/eu-vocabularies/concept-scheme/-/resource?uri=http://publications.europa.eu/resource/authority/corporate-body . Concepts referring to EP political groups will have 'EP_GROUP' as a Corporate body classification ",
                "required": false,
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "CD",
                            "C",
                            "CDI",
                            "COM",
                            "DE",
                            "DEP",
                            "CTDI",
                            "ADLE",
                            "ALDE",
                            "DR",
                            "ARC",
                            "ARE",
                            "EDD",
                            "EDN",
                            "ED",
                            "ECR",
                            "GUE/NGL",
                            "ENF",
                            "S&D",
                            "TDI",
                            "UPE",
                            "UEN",
                            "ITS",
                            "LDR",
                            "IND/DEM",
                            "L",
                            "NI",
                            "PSE",
                            "RDE",
                            "The Left",
                            "EFD",
                            "EFDD",
                            "EPP",
                            "PPE",
                            "EPP-ED",
                            "PPE-DE",
                            "ELDR",
                            "CG",
                            "FE",
                            "V",
                            "Greens/EFA",
                            "Verts/ALE",
                            "GUE",
                            "I-EDN",
                            "ID"
                        ]
                    }
                }
            },
...
            "parliamentary-term": {
                "name": "parliamentary-term",
                "in": "query",
                "description": "Parliamentary Term from 0 to the current.",
                "required": true,
                "style": "form",
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5,
                            6,
                            7,
                            8,
                            9
                        ],
                        "default": 9
                    }
                }
            },
...
            "mep-id": {
                "name": "mep-id",
                "in": "path",
                "description": "MEP identifier.",
                "required": true,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "default": 124936
                    }
                }
            },
...
            "offset": {
                "name": "offset",
                "in": "query",
                "description": "The number of items to skip before starting to collect the result set.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 0
                }
            },
            "limit": {
                "name": "limit",
                "in": "query",
                "description": "The number of items to return.",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
...
            "json-layout-fr": {
                "name": "json-layout",
                "in": "query",
                "description": "Define the layaout for Json+ld",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "framed"
                    ],
                    "default": "framed"
                }
            }
        }
    },
    "paths": {
        "/meps": {
            "get": {
                "tags": [
                    "MEPS"
                ],
                "summary": "Returns the list of all the MEPs",
                "description": "The service returns data about the MEPs related to the European Parliament.",
                "operationId": "get_meps",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/parliamentary-term"
                    },
                    {
                        "$ref": "#/components/parameters/gender"
                    },
                    {
                        "$ref": "#/components/parameters/political-group"
                    },
                    {
                        "$ref": "#/components/parameters/country-of-representation"
                    },
                    {
                        "$ref": "#/components/parameters/format-all"
                    },
                    {
                        "$ref": "#/components/parameters/offset"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MEPS"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/meps/{mep-id}": {
            "get": {
                "tags": [
                    "MEPS"
                ],
                "summary": "Returns a single MEP for the specified mep ID",
                "description": "The service returns a single MEP for the specified mep ID.",
                "operationId": "get_mep_by_id",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/mep-id"
                    },
                    {
                        "$ref": "#/components/parameters/format-all"
                    },
                    {
                        "$ref": "#/components/parameters/json-layout-fr"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}